fix : No Future Rewards For SmartCab
Future reward or gamma learning is not provided for the smartcab.

Smartcab only takes action based on the current state. Future state at
this step will change by the time smartcab goes to the next step.

Environment in which smart cab dirves is dymanic.So,Future rewards are
not provided to the SmartCab.